[LE-1762]: API for datadog test button for workflow and service guard (#8088)
* basic working datadog node-data API for logs

* rename getMetricsDataFor to getDataForNode since it now supports log data also for datadog

* integration test fix

* formatting

* set totalHits and totalHitsThreshold to null if not set
* it was getting set to zero

* test for datadog testbutton API

* support for workflow and 24-7
* hostname and hostNameField are optional now

* CustomLogDelegateService for fetching datadog logs

* Revert "CustomLogDelegateService for fetching datadog logs"

This reverts commit 0c74d3924e5274ae9503210ff5ee3ddca846bd61.

* create seperate APMValidateCollectorConfig for datadog logs
* datadog API does not need from and to parameters

* refactoring
* use better name and static variable convention for log API path suffix
* make method private

* typo fix

+++ 71-rest/src/test/java/software/wings/service/impl/elk/ElkAnalysisServiceImplTest.java
@@ -75,9 +75,9 @@ public class ElkAnalysisServiceImplTest extends WingsBaseTest {
 
     VerificationNodeDataSetupResponse verificationNodeDataSetupResponse =
         service.getLogDataByHost(accountId, elkSetupTestNodeData);
-    assertEquals(201, verificationNodeDataSetupResponse.getLoadResponse().getTotalHits());
+    assertEquals(201, (long) verificationNodeDataSetupResponse.getLoadResponse().getTotalHits());
     assertEquals(VerificationConstants.TOTAL_HITS_PER_MIN_THRESHOLD,
-        verificationNodeDataSetupResponse.getLoadResponse().getTotalHitsThreshold());
+        (long) verificationNodeDataSetupResponse.getLoadResponse().getTotalHitsThreshold());
   }
 
   @Test
